services/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = "Backend for Fitness App"
authors = ["reinhud <lrdev+github@proton.me>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
loguru = "^0.7.2"
sqlmodel = "^0.0.16"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.13.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.29.0"
asgi-lifespan = "^2.1.0"
sqlalchemy-utils = "^0.41.2"
nest-asyncio = "^1.6.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pre-commit = "^3.7.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
